refactor(database): move GetNeighbors/Walk/ShortestPath/ReadGraph to graph.go; leave stubs in database.go
